not clean code but it works
