+++ client/src/App.css
@@ -43,7 +43,7 @@
 }
 
 .feedback-form{
-  background-color: #e0e0e0;
+  background-color: #d7ddff;
   border-radius: 10px;
   width: 44vw;
   margin-left: 20vw;
@@ -62,12 +62,15 @@
 .feedback-buttons-row{
   width: 44vw;
   padding: 0.5em;
-  border: 1px solid black;
+  /* text-align: center; */
+  /* border: 1px solid black; */
+  display: flex;
+  justify-content: space-evenly;
 }
 
 .feedback-instructions{
   padding: 1em;
-  border: 1px solid black;
+  /* border: 1px solid black; */
   display: flex;
   justify-content: center;
   width: 44vw;
@@ -77,14 +80,14 @@
   width: 40vw;
   margin-left: 2vw;
   margin-bottom: 2vw;
-  border: 1px solid black;
-  border-radius: 10px;
-  background-color: #e0e0e0;
+  /* border: 1px solid black; */
+  border-radius: 5px;
+  height: 10em;
+  background-color: #ffffff;
   font-family: Arial, Helvetica, sans-serif;
 }
 
 .feedbackBtn{
-  display: flex;
-  justify-content: center;
-  margin: -2em;
+  text-align: center;
+  /* border: 1px solid black; */
 }
